devtools/client/webconsole/new-console-output/test/chrome/test_render_perf.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for getRepeatId()</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for render perf</p>
<div id="output"></div>

<script type="text/javascript">
"use strict";
const numMessages = 2000;
const testPackets = Array.from({length: numMessages}).map((el, id) => ({
  "from": "server1.conn4.child1/consoleActor2",
  "type": "consoleAPICall",
  "message": {
    "arguments": [
      "foobar",
      `${id % 2 === 0 ? "Even" : "Odd"} text`,
      id
    ],
    "columnNumber": 1,
    "counter": null,
    "filename": "file:///test.html",
    "functionName": "",
    "groupName": "",
    "level": "log",
    "lineNumber": 1,
    "private": false,
    "styles": [],
    "timeStamp": 1455064271115 + id,
    "timer": null,
    "workerType": "none",
    "category": "webdev"
  }
}));

async function timeit(cb) {
  // Return a Promise that resolves the number of seconds cb takes.
  let start = performance.now();
  await cb();
  let elapsed = performance.now() - start;
  return elapsed;
}

window.onload = async function () {
  try {
    const NewConsoleOutputWrapper = browserRequire(
      "devtools/client/webconsole/new-console-output/new-console-output-wrapper");
    const EventEmitter = browserRequire("devtools/shared/event-emitter");

    const wrapper = new NewConsoleOutputWrapper(
      document.getElementById("output"),
      {hud: EventEmitter.decorate({proxy: {}})},
      {},
      null,
      document,
    );
    wrapper.init();

    let times = [];
    const iterations = 10;
    const lastPacket = testPackets.pop();
    for (let i = 0; i < iterations; i++) {
      let time = await timeit(() => {
        testPackets.forEach((packet) => wrapper.dispatchMessageAdd(packet));
        // Only wait for the last packet to minimize work.
        return wrapper.dispatchMessageAdd(lastPacket, true);
      });
      info(`took ${time} ms to render messages`);
      times.push(time);

      info("Clear the console");
      await new Promise(resolve => requestAnimationFrame(() => resolve()));
      wrapper.dispatchMessagesClear();
    }

    let totalTime = times.reduce((sum, t) => sum + t);
    let avg = totalTime / times.length;
    info(`On average, it took ${avg} ms to render ${numMessages} messages`);

    ok(true, "Yay, it didn't time out!");
  } catch (e) {
    ok(false, `Error :  ${e.message}
      ${e.stack}
    `);
  }

  SimpleTest.finish();
};
</script>
</body>
</html>
